Tools/Vectorization/requirements.txt
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
beautifulsoup4>=4.9.3,<5.0.0
openai>=1.3.0,<2.0.0
tiktoken>=0.3.0,<1.0.0
colorama>=0.4.4,<1.0.0
azure-core>=1.26.1,<2.0.0
azure-search-documents>=11.5.0,<12.0.0